requirements.txt
altair==4.2.0
imageio==2.16.1
numpy==1.21.5
opencv-python-headless==4.5.5.62
pandas==1.3.5
Pillow==9.0.1
pydub==0.25.1
streamlit==1.3.1
streamlit-webrtc==0.35.1
pytube3==12.0.0
# ffmpeg-python==0.2.0
# future==0.18.2
# ffprobe-python==1.0.3
